[rename]: Moved AccountPage, DashboardPage, and ErrorPage
- moved AccountPage to Account folder
- moved DashboardPage to Dashboard
- moved ErrorPage to util
