+++ frontend/src/pages/DashBoard/DashboardPage.js
@@ -2,16 +2,16 @@ import { Box } from '@chakra-ui/react';
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import useAxios from '../hooks/useAxios';
-import useFormatAmount from '../hooks/useFormatAmount';
-import Container from '../components/container/Container';
-import Icon from '../components/util/Icon';
-import CustomText from '../components/CustomText';
-import InfoBlock from '../components/util/InfoBlock';
-import DateTimeDisplay from '../components/util/DateTimeDisplay';
-import ContainerRowBalanceWrapper from '../components/container/ContainerRowBalanceWrapper';
-import CustomBox from '../components/util/CustomBox';
-import Calendar from '../components/calendar/Calendar';
+import useAxios from '../../hooks/useAxios';
+import useFormatAmount from '../../hooks/useFormatAmount';
+import Container from '../../components/container/Container';
+import Icon from '../../components/util/Icon';
+import CustomText from '../../components/CustomText';
+import InfoBlock from '../../components/util/InfoBlock';
+import DateTimeDisplay from '../../components/util/DateTimeDisplay';
+import ContainerRowBalanceWrapper from '../../components/container/ContainerRowBalanceWrapper';
+import CustomBox from '../../components/util/CustomBox';
+import Calendar from '../../components/calendar/Calendar';
 
 
 const MONTHS = {
