+++ frontend/src/utils/AppRouter.js
@@ -9,9 +9,9 @@ const Home = lazy(() => import('../pages/Home'));
 const Login = lazy(() => import('../pages/Login'));
 const Signup = lazy(() => import('../pages/Signup'));
 const Kyc = lazy(() => import('../pages/KYC'));
-const DashBoard = lazy(() => import('../pages/DashboardPage'));
+const DashBoard = lazy(() => import('../pages/DashBoard/DashboardPage'));
 const NotFound404 = lazy(() => import('../pages/ErrorPage'));
-const Account = lazy(() => import('../pages/AccountPage'));
+const Account = lazy(() => import('../pages/Account/AccountPage'));
 // Remittance
 const Remittance = lazy(() => import('../pages/Remittance/Remittance'));
 const Transfer = lazy(() => import('../pages/Remittance/Transfer'));
